Refactored function callback
